+++ packages/p2p/source/service-provider.ts
@@ -3,6 +3,10 @@ import { Providers, Services, Utils } from "@mainsail/kernel";
 import Joi from "joi";
 
 import { RevalidateApiNodeAction, ValidateAndAcceptApiNodeAction, ValidateAndAcceptPeerAction } from "./actions";
+import { ApiNodeDiscoverer } from "./api-node-discoverer";
+import { ApiNodeProcessor } from "./api-node-processor";
+import { ApiNode, ApiNodeRepository } from "./api-node-repository";
+import { ApiNodeVerifier } from "./api-node-verifier";
 import { Broadcaster } from "./broadcaster";
 import { BlockDownloader } from "./downloader/block-downloader";
 import { MessageDownloader } from "./downloader/message-downloader";
@@ -11,10 +15,6 @@ import { Header } from "./header";
 import { HeaderService } from "./header-service";
 import { Logger } from "./logger";
 import { Peer } from "./peer";
-import { ApiNodeDiscoverer } from "./peer-api-node-discoverer";
-import { ApiNodeProcessor } from "./peer-api-node-processor";
-import { ApiNode, ApiNodeRepository } from "./peer-api-node-repository";
-import { ApiNodeVerifier } from "./peer-api-node-verifier";
 import { PeerCommunicator } from "./peer-communicator";
 import { PeerConnector } from "./peer-connector";
 import { PeerDiscoverer } from "./peer-discoverer";
